Split and refactor tree view data tests
